+++ CHANGELOG.md
@@ -1,15 +1,16 @@
 CHANGELOG
 =========
 
-[Next release](https://github.com/rebing/graphql-laravel/compare/6.3.0...master)
+[Next release](https://github.com/rebing/graphql-laravel/compare/6.4.0...master)
 --------------
 
+2021-03-31, 6.4.0
+-----------------
 ### Added
 - Support for per-schema types [\#658 / stevelacey](https://github.com/rebing/graphql-laravel/pull/658)
 
 2021-03-12, 6.3.0
 -----------------
-
 ### Added
 - Support for class based schemas [\#706 / jasonvarga](https://github.com/rebing/graphql-laravel/pull/706)